+++ api/src/main/java/org/ehrbase/api/dto/AqlQueryContext.java
@@ -29,6 +29,9 @@ public interface AqlQueryContext {
     enum EhrbaseMetaProperty implements MetaProperty {
         OFFSET("offset"),
         FETCH("fetch"),
+        DEFAULT_LIMIT("default-limit"),
+        MAX_LIMIT("max-limit"),
+        MAX_FETCH("max-fetch"),
         RESULT_SIZE("resultsize"),
         DRY_RUN("dry_run"),
         EXECUTED_SQL("executed_sql"),