+++ aql-engine/src/main/java/org/ehrbase/openehr/aqlengine/service/AqlQueryServiceImp.java
@@ -35,6 +35,7 @@ import java.util.Optional;
 import java.util.stream.LongStream;
 import java.util.stream.Stream;
 import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang3.ObjectUtils;
 import org.ehrbase.api.dto.AqlQueryContext;
 import org.ehrbase.api.dto.AqlQueryRequest;
 import org.ehrbase.api.exception.AqlFeatureNotImplementedException;
@@ -88,13 +89,55 @@ public class AqlQueryServiceImp implements AqlQueryService {
     private final AqlQueryContext aqlQueryContext;
 
     @Value("${ehrbase.rest.aql.default-limit:}")
-    private Optional<Long> defaultLimit = Optional.empty();
+    private Long defaultLimit;
 
     @Value("${ehrbase.rest.aql.max-limit:}")
-    private Optional<Long> maxLimit = Optional.empty();
+    private Long maxLimit;
 
     @Value("${ehrbase.rest.aql.max-fetch:}")
-    private Optional<Long> maxFetch = Optional.empty();
+    private Long maxFetch;
+
+    enum FetchPrecedence {
+        /**
+         * Fail if both fetch and limit are present
+         */
+        REJECT,
+        /**
+         * Take minimum of fetch and limit for limit;
+         * fail if query has offset
+         */
+        MIN_FETCH;
+    }
+
+    @Value("${ehrbase.rest.aql.fetch-precedence:REJECT}")
+    private FetchPrecedence fetchPrecedence = FetchPrecedence.REJECT;
+
+    private static Long applyFetchPrecedence(FetchPrecedence fetchPrecedence, Long queryLimit, Long queryOffset, Long fetchParam, Long offsetParam) {
+        if (fetchParam == null) {
+            if (offsetParam != null) {
+                throw new UnprocessableEntityException(
+                        "Query parameter for offset provided, but no fetch parameter");
+            }
+            return queryLimit;
+        } else if (queryLimit == null) {
+            assert queryOffset == null;
+            return fetchParam;
+        }
+
+        return switch (fetchPrecedence) {
+            case REJECT -> {
+                    throw new UnprocessableEntityException(
+                            "Query contains a LIMIT clause, fetch and offset parameters must not be used (with fetch precedence %s)".formatted(fetchPrecedence));
+            }
+            case MIN_FETCH -> {
+                if (queryOffset != null) {
+                    throw new UnprocessableEntityException(
+                            "Query contains a OFFSET clause, fetch parameter must not be used (with fetch precedence %s)".formatted(fetchPrecedence));
+                }
+                yield Math.min(queryLimit, fetchParam);
+            }
+        };
+    }
 
     @Autowired
     public AqlQueryServiceImp(
@@ -119,16 +162,22 @@ public class AqlQueryServiceImp implements AqlQueryService {
 
     private QueryResultDto queryAql(AqlQueryRequest aqlQueryRequest) {
 
-        defaultLimit.ifPresent(v -> aqlQueryContext.setMetaProperty(
-                AqlQueryContext.EhrbaseMetaProperty.DEFAULT_LIMIT, v));
-        maxLimit.ifPresent(v -> aqlQueryContext.setMetaProperty(
-                AqlQueryContext.EhrbaseMetaProperty.MAX_LIMIT, v));
-        maxFetch.ifPresent(v -> aqlQueryContext.setMetaProperty(
-                AqlQueryContext.EhrbaseMetaProperty.MAX_FETCH, v));
+        if (defaultLimit != null) {
+            aqlQueryContext.setMetaProperty(
+                    AqlQueryContext.EhrbaseMetaProperty.DEFAULT_LIMIT, defaultLimit);
+        }
+        if (maxLimit != null) {
+            aqlQueryContext.setMetaProperty(
+                    AqlQueryContext.EhrbaseMetaProperty.MAX_LIMIT, maxLimit);
+        }
+        if (maxFetch != null) {
+            aqlQueryContext.setMetaProperty(
+                    AqlQueryContext.EhrbaseMetaProperty.MAX_FETCH, maxFetch);
+        }
 
         // TODO: check that select aliases are not duplicated
         try {
-            AqlQuery aqlQuery = buildAqlQuery(aqlQueryRequest, defaultLimit, maxLimit, maxFetch);
+            AqlQuery aqlQuery = buildAqlQuery(aqlQueryRequest, fetchPrecedence, defaultLimit, maxLimit, maxFetch);
 
             aqlQueryFeatureCheck.ensureQuerySupported(aqlQuery);
 
@@ -203,62 +252,35 @@ public class AqlQueryServiceImp implements AqlQueryService {
 
     static AqlQuery buildAqlQuery(
             AqlQueryRequest aqlQueryRequest,
-            Optional<Long> defaultLimit,
-            Optional<Long> maxLimit,
-            Optional<Long> maxFetch) {
+            FetchPrecedence fetchPrecedence,
+            Long defaultLimit,
+            Long maxLimit,
+            Long maxFetch) {
 
         AqlQuery aqlQuery = AqlQueryParser.parse(aqlQueryRequest.queryString());
 
         // apply limit and offset - where the definitions from the aql are the precedence
         Optional<AqlQueryRequest> qr = Optional.of(aqlQueryRequest);
-        Optional<Long> fetchParam = qr.map(AqlQueryRequest::fetch);
-        Optional<Long> offsetParam = qr.map(AqlQueryRequest::offset);
+        Long fetchParam = aqlQueryRequest.fetch();
+        Long offsetParam = aqlQueryRequest.offset();
 
-        Optional<AqlQuery> q = Optional.of(aqlQuery);
-        Optional<Long> limit = q.map(AqlQuery::getLimit);
-        Optional<Long> offset = q.map(AqlQuery::getOffset);
+        Long queryLimit = aqlQuery.getLimit();
+        Long queryOffset = aqlQuery.getOffset();
 
-        // verify not parameter fetch offset are defined when query contains a LIMIT or assign fetch parameter
-        if (limit.isPresent()) {
-            limit.flatMap(l -> maxLimit.filter(m -> l > m)
-                            .map(m -> new UnprocessableEntityException(
-                                    "Query LIMIT %d exceeds maximum limit %d".formatted(l, m))))
-                    .ifPresent(ex -> {
-                        throw ex;
-                    });
-
-            if (fetchParam.isPresent() || offsetParam.isPresent()) {
-                throw new UnprocessableEntityException(
-                        "Query contains a LIMIT clause, fetch and offset parameters must not be used");
-            }
-
-        } else {
-            if (fetchParam.isPresent()) {
-                // set from parameters
-                fetchParam
-                        .flatMap(l -> maxFetch.filter(m -> l > m)
-                                .map(m -> new UnprocessableEntityException(
-                                        "Fetch parameter %d exceeds maximum fetch %d".formatted(l, m))))
-                        .ifPresent(ex -> {
-                            throw ex;
-                        });
-
-                limit = fetchParam;
-                offset = offsetParam;
-
-            } else {
-                offsetParam.ifPresent(os -> {
-                    throw new UnprocessableEntityException(
-                            "Query parameter for offset provided, but no fetch parameter");
-                });
-
-                // set default
-                limit = defaultLimit;
-            }
-
-            aqlQuery.setLimit(limit.orElse(null));
-            aqlQuery.setOffset(offset.orElse(null));
+        if (queryLimit != null && maxLimit != null && queryLimit > maxLimit) {
+            throw  new UnprocessableEntityException(
+                    "Query LIMIT %d exceeds maximum limit %d".formatted(queryLimit, maxLimit));
         }
+
+        if (fetchParam != null && maxFetch != null && fetchParam > maxFetch) {
+            throw  new UnprocessableEntityException(
+                    "Fetch parameter %d exceeds maximum fetch %d".formatted(fetchParam, maxFetch));
+        }
+
+        Long limit = applyFetchPrecedence(fetchPrecedence, queryLimit, queryOffset, fetchParam, offsetParam);
+
+        aqlQuery.setLimit(ObjectUtils.firstNonNull(limit, defaultLimit));
+        aqlQuery.setOffset(ObjectUtils.firstNonNull(offsetParam, queryOffset));
 
         // postprocess
         replaceParameters(aqlQuery, aqlQueryRequest.parameters());